backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
groq>=0.4.1
httpx>=0.25.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.23
pg8000>=1.30.0
numpy>=1.26.0
sentence-transformers>=3.0.1
PyPDF2>=3.0.1
aiofiles>=23.2.1
email-validator>=2.1.0